Add Nexus ready systemd notification option
When the Nexus Server is run by the Compositor Plugin and the WPE
Framework service is brought up via systemd, it is important that the
Nexus Server notifies systemd when it is ready for accepting
connections, so that systemd can wait for this notification before
starting other services that need to connect to Nexus. This is a feature
that is implemented in Broadcom's reference 'nxserver' service (when
it's compiled with the 'ENABLE_SD_NOTIFY' option).

This way, add a new 'WPEFRAMEWORK_NEXUS_SERVER_SD_NOTIFY' option to the
compositor plugin (disabled by default) which, when enabled, performs
the mentioned systemd notification, by using the systemd's library
'sd_notifyf()' function. The notification is done in the new
'Platform::NotifyReady()' method and called right after
'nxserver_ipc_init()' succeeds from 'Platform::Platform()'.

With the new option enabled, there's a new dependency to the systemd
library, so add a new cmake find module ('FindSystemd.cmake') for
making sure the dependency can be satisfied and get the required
elements (includes, libs and definitions).

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>