[IOConnector] Fix use of invalid identifier in IOConnector::Resource(uint32_t id)
The current implementation of IExternal::IFactory::Resource(uint32_t id)
in IOConnector plugin checks the "id" with the GPIO::Pin::Identifier()
for all the GPIO pins we have configured and return pointer to GPIO::Pin
object which matches our "id".

But we need to mask the identifier returned by Identifier() method
before use, This is because when we are creating GPIO::Pin objects the
GPIO pin id is masked with a prefix corresponding to the type of
ExternalBase<> template we are creating.

In our case "class Pin" is created from ExternalBase of type
Exchange::IExternal::GPIO and this our gpio pin id will be masked with
IExternal::identification::GPIO which is 0x20000000. See
Interfaces/IExternal.h for more details.

Thus mask the GPIO::Pin::Identifier() before comparison, We have used
the mask as 0xFFFF and not 0x0FFFFFFF following other places in
IOConnector where it is already used.

Signed-off-by: harikrishnan.p <harikrishnan.p@tataelxsi.co.in>